refactor(cli): extract fullscreen composer module
