+++ packages/agent-cli/agent-cli.cabal
@@ -83,8 +83,10 @@ library
                     , Agent.CLI.Tools
                     , Agent.CLI.TUI.App
                     , Agent.CLI.TUI.Bridge
+                    , Agent.CLI.TUI.Composer
                     , Agent.CLI.TUI.ImagePreview
                     , Agent.CLI.TUI.Scroll
+                    , Agent.CLI.TUI.Types
                     , Agent.CLI.Turn
                     , Agent.CLI.Usage
                     , Agent.CLI.Worktree
@@ -175,6 +177,7 @@ test-suite agent-cli-test
                     , Agent.CLI.ToolsSpec
                     , Agent.CLI.TUIAppSpec
                     , Agent.CLI.TUIBridgeSpec
+                    , Agent.CLI.TUIComposerSpec
                     , Agent.CLI.TUIImagePreviewSpec
                     , Agent.CLI.TUIScrollSpec
                     , Agent.CLI.UsageSpec
@@ -198,6 +201,7 @@ test-suite agent-cli-test
                     , JuicyPixels
                     , process
                     , safe-exceptions
+                    , stm
                     , text
                     , time
                     , transformers
