+++ packages/agent-cli/test/Spec.hs
@@ -46,6 +46,7 @@ import qualified Agent.CLI.TerminalSpec as TerminalSpec
 import qualified Agent.CLI.ToolsSpec as ToolsSpec
 import qualified Agent.CLI.TUIAppSpec as TUIAppSpec
 import qualified Agent.CLI.TUIBridgeSpec as TUIBridgeSpec
+import qualified Agent.CLI.TUIComposerSpec as TUIComposerSpec
 import qualified Agent.CLI.TUIImagePreviewSpec as TUIImagePreviewSpec
 import qualified Agent.CLI.TUIScrollSpec as TUIScrollSpec
 import qualified Agent.CLI.UsageSpec as UsageSpec
@@ -97,6 +98,7 @@ main = hspec do
     ToolsSpec.spec
     TUIAppSpec.spec
     TUIBridgeSpec.spec
+    TUIComposerSpec.spec
     TUIImagePreviewSpec.spec
     TUIScrollSpec.spec
     UsageSpec.spec
